refactor of file naming and package build to allow test without build
