+++ test/index.test.mjs
@@ -2,10 +2,13 @@
 import chai from 'chai';
 import { ethers } from 'ethers';
 import fetch from 'node-fetch';
-import elasticSwapSDK from '../dist/index.js';
+import * as elasticSwapSDK from '../src/index.mjs';
+import LocalStorageAdapterMock from './adapters/LocalStorageAdapterMock.mjs';
 
 const { assert } = chai;
 const RPC_URL = 'https://mainnet.infura.io/v3/48f877fa4aa4490bb0c988368dc8e373';
+
+const storageAdapter = new LocalStorageAdapterMock();
 
 describe('SDK', () => {
   const env = {
@@ -16,13 +19,13 @@ describe('SDK', () => {
   describe('Constructor', () => {
     it('Can be created via constructor', async () => {
       const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
-      const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider });
+      const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider, storageAdapter });
       assert.isNotNull(sdk);
     });
 
     it('Sets the block number', async () => {
       const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
-      const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider });
+      const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider, storageAdapter });
 
       await provider.getBlockNumber();
 
@@ -32,7 +35,7 @@ describe('SDK', () => {
 
     it('Creates the ExchangeFactory', async () => {
       const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
-      const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider });
+      const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider, storageAdapter });
       assert.isNotNull(sdk.exchangeFactory);
     });
   });
@@ -40,18 +43,18 @@ describe('SDK', () => {
   describe('setName', () => {
     it('sets name correctly with address that has no ENS name', async () => {
       const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
-      let sdk = new elasticSwapSDK.SDK({ account: '0xC9e4c8a2F2D8D684fB9de60aBFE3Fb5Ea7565366', env, customFetch: fetch, provider });
+      let sdk = new elasticSwapSDK.SDK({ account: '0xC9e4c8a2F2D8D684fB9de60aBFE3Fb5Ea7565366', env, customFetch: fetch, provider, storageAdapter });
       await sdk.setName();
       assert.equal('0xC9e4...5366', sdk.name);
 
-      sdk = new elasticSwapSDK.SDK({ account: '0xE16584424F34380DBf798f547Af684597788FbC7', env, customFetch: fetch, provider });
+      sdk = new elasticSwapSDK.SDK({ account: '0xE16584424F34380DBf798f547Af684597788FbC7', env, customFetch: fetch, provider, storageAdapter });
       await sdk.setName();
       assert.equal('0xean.eth', sdk.name);
     });
 
     it('sets name correctly with address that has an ENS name', async () => {
       const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
-      const sdk = new elasticSwapSDK.SDK({ account: '0xE16584424F34380DBf798f547Af684597788FbC7', env, customFetch: fetch, provider });
+      const sdk = new elasticSwapSDK.SDK({ account: '0xE16584424F34380DBf798f547Af684597788FbC7', env, customFetch: fetch, provider, storageAdapter });
       await sdk.setName();
       assert.equal('0xean.eth', sdk.name);
     });
@@ -59,7 +62,7 @@ describe('SDK', () => {
 
   describe('isValidETHAddress', () => {
     const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
-    const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider });
+    const sdk = new elasticSwapSDK.SDK({ env, customFetch: fetch, provider, storageAdapter });
 
     it('returns true for a valid address', async () => {
       const isValid = await sdk.isValidETHAddress('0xC9e4c8a2F2D8D684fB9de60aBFE3Fb5Ea7565366');
